requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
aiofiles=22.1.0=pypi_0
aiohttp=3.9.5=pypi_0
aiosignal=1.3.1=pypi_0
aiosqlite=0.20.0=pypi_0
altair=5.3.0=pypi_0
analytics-python=1.4.post1=pypi_0
annotated-types=0.6.0=pypi_0
anyio=4.3.0=pypi_0
apispec=6.6.1=pypi_0
appnope=0.1.4=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.3.0=pypi_0
asttokens=2.4.1=pypi_0
astunparse=1.6.3=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
babel=2.14.0=pypi_0
backoff=1.10.0=pypi_0
beautifulsoup4=4.12.3=pypi_0
bech32=1.2.0=pypi_0
bleach=6.1.0=pypi_0
blinker=1.8.1=pypi_0
bokeh=3.4.1=pypi_0
bqplot=0.12.43=pypi_0
branca=0.7.2=pypi_0
ca-certificates=2024.3.11=hecd8cb5_0
cachetools=5.3.3=pypi_0
certifi=2024.2.2=pypi_0
cffi=1.16.0=pypi_0
chardet=5.2.0=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
colour=0.1.5=pypi_0
comm=0.2.2=pypi_0
contourpy=1.2.1=pypi_0
cosmpy=0.9.2=pypi_0
crewai=0.5.0=pypi_0
cycler=0.12.1=pypi_0
dataclasses-json=0.6.4=pypi_0
debugpy=1.8.1=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
diskcache=5.6.3=pypi_0
distro=1.9.0=pypi_0
docutils=0.21.2=pypi_0
duckdb=0.10.2=pypi_0
ecdsa=0.18.0=pypi_0
entrypoints=0.4=pypi_0
et-xmlfile=1.1.0=pypi_0
exceptiongroup=1.2.1=pypi_0
executing=2.0.1=pypi_0
fastjsonschema=2.19.1=pypi_0
filelock=3.14.0=pypi_0
flaml=2.1.2=pypi_0
folium=0.16.0=pypi_0
fonttools=4.51.0=pypi_0
fqdn=1.5.1=pypi_0
frozenlist=1.4.1=pypi_0
gdown=5.1.0=pypi_0
geojson=3.1.0=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
googleapis-common-protos=1.63.0=pypi_0
greenlet=3.0.3=pypi_0
grpcio=1.62.2=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.5=pypi_0
httpx=0.27.0=pypi_0
idna=3.7=pypi_0
importlib-metadata=7.1.0=pypi_0
importlib-resources=6.4.0=pypi_0
ipyevents=2.0.2=pypi_0
ipyfilechooser=0.6.0=pypi_0
ipykernel=6.29.4=pypi_0
ipyleaflet=0.19.0=pypi_0
ipython=8.18.1=pypi_0
ipython-genutils=0.2.0=pypi_0
ipytree=0.2.2=pypi_0
ipywidgets=8.1.2=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.3=pypi_0
json5=0.9.25=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=2.4=pypi_0
jsonschema=4.21.1=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter-client=7.4.9=pypi_0
jupyter-core=5.7.2=pypi_0
jupyter-events=0.10.0=pypi_0
jupyter-leaflet=0.19.0=pypi_0
jupyter-server=2.14.0=pypi_0
jupyter-server-fileid=0.9.2=pypi_0
jupyter-server-terminals=0.5.3=pypi_0
jupyter-server-ydoc=0.8.0=pypi_0
jupyter-ydoc=0.2.5=pypi_0
jupyterlab=3.6.7=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab-server=2.27.1=pypi_0
jupyterlab-widgets=3.0.10=pypi_0
kiwisolver=1.4.5=pypi_0
krb5=1.20.1=h428f121_1
langchain=0.1.0=pypi_0
langchain-community=0.0.20=pypi_0
langchain-core=0.1.23=pypi_0
langchain-openai=0.0.2.post1=pypi_0
langchain-text-splitters=0.0.1=pypi_0
langsmith=0.0.87=pypi_0
leafmap=0.31.9=pypi_0
libcxx=14.0.6=h9765a3e_0
libedit=3.1.20230828=h6c40b1e_0
libffi=3.4.4=hecd8cb5_0
libpq=12.17=h04015c4_0
llamaapi=0.1.36=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
marshmallow=3.21.1=pypi_0
matplotlib=3.8.4=pypi_0
matplotlib-inline=0.1.7=pypi_0
mdurl=0.1.2=pypi_0
mistune=3.0.2=pypi_0
mitosheet=0.1.573=pypi_0
monotonic=1.6=pypi_0
msgpack=1.0.8=pypi_0
multidict=6.0.5=pypi_0
mypy-extensions=1.0.0=pypi_0
nbclassic=1.0.0=pypi_0
nbclient=0.10.0=pypi_0
nbconvert=7.16.4=pypi_0
nbformat=5.10.4=pypi_0
ncurses=6.4=hcec6c5f_0
nest-asyncio=1.6.0=pypi_0
notebook=6.5.6=pypi_0
notebook-shim=0.2.4=pypi_0
numpy=1.26.4=pypi_0
openai=0.28.1=pypi_0
openpyxl=3.1.2=pypi_0
openssl=3.0.13=hca72f7f_0
orjson=3.10.1=pypi_0
overrides=7.7.0=pypi_0
packaging=23.2=pypi_0
pages=0.3=pypi_0
pandas=2.2.2=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.4=pypi_0
pexpect=4.9.0=pypi_0
piglet-templates=1.3.1=pypi_0
pillow=10.3.0=pypi_0
pip=23.3.1=py39hecd8cb5_0
platformdirs=4.2.1=pypi_0
plotly=5.21.0=pypi_0
prometheus-client=0.20.0=pypi_0
prompt-toolkit=3.0.43=pypi_0
protobuf=4.25.3=pypi_0
psutil=5.9.8=pypi_0
psycopg2=2.9.9=py39h6c40b1e_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyarrow=16.0.0=pypi_0
pyautogen=0.1.14=pypi_0
pycparser=2.22=pypi_0
pycryptodome=3.20.0=pypi_0
pydantic=2.7.1=pypi_0
pydantic-core=2.18.2=pypi_0
pydeck=0.9.0b1=pypi_0
pygments=2.17.2=pypi_0
pyparsing=3.1.2=pypi_0
pyshp=2.3.1=pypi_0
pysocks=1.7.1=pypi_0
pystac=1.10.0=pypi_0
pystac-client=0.7.7=pypi_0
python=3.9.19=h5ee71fb_0
python-box=7.1.1=pypi_0
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
python-json-logger=2.0.7=pypi_0
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=24.0.1=pypi_0
readline=8.2=hca72f7f_0
referencing=0.35.0=pypi_0
regex=2024.4.28=pypi_0
requests=2.31.0=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rich=13.7.1=pypi_0
rpds-py=0.18.0=pypi_0
scooby=0.9.2=pypi_0
send2trash=1.8.3=pypi_0
setuptools=68.2.2=py39hecd8cb5_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.5=pypi_0
sqlalchemy=2.0.29=pypi_0
sqlite=3.41.2=h6c40b1e_0
stack-data=0.6.3=pypi_0
state=0.1.1.dev0=pypi_0
streamlit=1.33.0=pypi_0
streamlit-bokeh-events=0.1.2=pypi_0
streamlit-geolocation=0.0.10=pypi_0
streamlit-image-select=0.6.0=pypi_0
tenacity=8.2.3=pypi_0
termcolor=2.4.0=pypi_0
terminado=0.18.1=pypi_0
tiktoken=0.5.2=pypi_0
tinycss2=1.3.0=pypi_0
tk=8.6.12=h5d9f67b_0
toml=0.10.2=pypi_0
tomli=2.0.1=pypi_0
toolz=0.12.1=pypi_0
tornado=6.4=pypi_0
tqdm=4.66.2=pypi_0
traitlets=5.14.3=pypi_0
traittypes=0.2.1=pypi_0
types-python-dateutil=2.9.0.20240316=pypi_0
typing-extensions=4.11.0=pypi_0
typing-inspect=0.9.0=pypi_0
tzdata=2024.1=pypi_0
uagents=0.11.1=pypi_0
uri-template=1.3.0=pypi_0
urllib3=2.2.1=pypi_0
uvicorn=0.20.0=pypi_0
wcwidth=0.2.13=pypi_0
webcolors=1.13=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.8.0=pypi_0
websockets=10.4=pypi_0
wheel=0.41.2=py39hecd8cb5_0
whitebox=2.3.1=pypi_0
whiteboxgui=2.3.0=pypi_0
widgetsnbextension=4.0.10=pypi_0
xyzservices=2024.4.0=pypi_0
xz=5.4.6=h6c40b1e_0
y-py=0.6.2=pypi_0
yarl=1.9.4=pypi_0
ypy-websocket=0.8.4=pypi_0
zipp=3.18.1=pypi_0
zlib=1.2.13=h4dc903c_0
